Rename `HciController` to `ExternalController` and move to the `controller` module
